pyproject.toml
[project]
name = "resume-optimizer-project"
version = "0.1.0"
description = ""
authors = [
    {name = "uttara-sriya",email = "sriya.uttara@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "google-adk (>=1.27.2,<2.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "python-docx (>=1.2.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
